backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
faster-whisper
langchain
langchain-openai
notion-client
python-dotenv
alembic
